Updated the max rate of flanger and max delay of echo. Fixed some code that somehow got reverted.

+++ Garble_Me_Project/src/main/java/com/codepotato/model/Player.java
@@ -123,7 +123,7 @@ public class Player implements Runnable{
      * from 0-100
      */
     public int currPositionPercent(){
-        return Math.round(sampleReader.getPosition() / sampleReader.length() * 100);
+        return (int)Math.round((double)sampleReader.getPosition() / (double)sampleReader.length() * 100.);
     }
 
 }
